# backend/src/main/resources/openapi/api.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "HEH Gourmet",
        "description": "API for the HEH cafeteria online project",
        "version": "0.0.1",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "contact": {
            "name": "HEH Gourmet Support"
        }
    },
    "servers": [
        {
            "description": "localhost dev server",
            "url": "https:localhost:8080"
        },
        {
            "description": "Production server",
            "url": "https://gourmet-iso4.onrender.com"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "OAuth": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {
                            "write": "Grants write access"
                        }
                    }
                }
            }
        },
...
        "schemas": {
...
            "Category": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "descriptor": {
                        "type": "string"
                    }
                }
            },
            "InputCategory": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "descriptor": {
                        "type": "string"
                    }
                }
            },
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "float"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "image": {
                        "type": "url"
                    },
                    "categoryID": {
                        "type": "integer"
                    }
                }
            },
...
        }
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get all categories",
                "operationId": "getCategories",
                "responses": {
                    "200": {
                        "description": "A list of category.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create a category in bulk",
                "operationId": "createCategories",
                "requestBody": {
                    "description": "Category to create",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/InputCategory"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Products created"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete all categories",
                "operationId": "deleteCategories",
                "responses": {
                    "200": {
                        "description": "All categories deleted"
                    }
                },
                "requestBody": {
                    "description": "Category to delete",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get a category by id",
                "operationId": "getCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A category.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Update a category",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Category to update",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InputCategory"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Category updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "summary": "Delete a category",
                "operationId": "deleteCategory",
                "responses": {
                    "200": {
                        "description": "Category deleted"
                    }
                }
            }
        },
...
        "/api/cart/{userID}": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get all products from a cart",
                "parameters": [
                    {
                        "name": "userID",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All products from a cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add a product to a cart",
                "parameters": [
                    {
                        "name": "userID",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "productID",
                        "in": "query",
                        "description": "ID of the product to add",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Quantity of the product to add",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product put in cart"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Delete a product from a cart",
                "parameters": [
                    {
                        "name": "userID",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "productID",
                        "in": "query",
                        "description": "ID of the product to add",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Quantity of the product to add",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted from cart"
                    }
                }
            }
        },
        "/api/cart/{userID}/checkout": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Checkout a cart",
                "parameters": [
                    {
                        "name": "userID",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart checked out"
                    }
                }
            }
        }
    }
}